Move toast into useRestore hook
